requirements.txt
opencv-python
numpy
mediapy
matplotlib
Pillow
wget
transformers
typing-extensions
tqdm
requests
ipykernel
iprogress
ipywidgets
importlib

umap-learn
scikit-learn